requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.3.1=pypi_0
accelerate=1.0.1=pypi_0
aiofiles=23.2.1=pypi_0
aiohappyeyeballs=2.6.1=pypi_0
aiohttp=3.12.15=pypi_0
aiosignal=1.4.0=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.10.0=pypi_0
async-timeout=5.0.1=pypi_0
attrs=25.3.0=pypi_0
bitsandbytes=0.45.0=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.9.9=h06a4308_0
certifi=2025.8.3=pypi_0
cffi=2.0.0=pypi_0
charset-normalizer=3.4.3=pypi_0
click=8.1.7=pypi_0
cmake=3.31.1=pypi_0
contourpy=1.1.1=pypi_0
cycler=0.12.1=pypi_0
diffusers=0.31.0=pypi_0
docker-pycreds=0.4.0=pypi_0
einops=0.8.0=pypi_0
eval-type-backport=0.2.0=pypi_0
exceptiongroup=1.3.0=pypi_0
fastapi=0.116.1=pypi_0
ffmpy=0.6.1=pypi_0
filelock=3.19.1=pypi_0
fire=0.7.0=pypi_0
flatbuffers=25.2.10=pypi_0
fonttools=4.55.1=pypi_0
frozenlist=1.7.0=pypi_0
fsspec=2025.9.0=pypi_0
ftfy=6.2.3=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
gradio=5.14.0=pypi_0
gradio-client=1.7.0=pypi_0
h11=0.16.0=pypi_0
httpcore=1.0.9=pypi_0
httpx=0.28.1=pypi_0
huggingface-hub=0.26.3=pypi_0
idna=3.10=pypi_0
imageio=2.37.0=pypi_0
importlib-metadata=8.5.0=pypi_0
importlib-resources=6.4.5=pypi_0
jax=0.4.30=pypi_0
jaxlib=0.4.30=pypi_0
jinja2=3.1.6=pypi_0
joblib=1.4.2=pypi_0
kiwisolver=1.4.7=pypi_0
lazy-loader=0.4=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
libxcb=1.17.0=h9b100fa_0
libzlib=1.3.1=hb25bd0a_0
lightning-utilities=0.9.0=pypi_0
lit=18.1.8=pypi_0
lora-diffusion=0.1.7=pypi_0
markdown=3.7=pypi_0
markdown-it-py=4.0.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.7.5=pypi_0
mdurl=0.1.2=pypi_0
mediapipe=0.10.11=pypi_0
ml-dtypes=0.5.3=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.6.4=pypi_0
ncurses=6.5=h7934f7d_0
networkx=3.4.2=pypi_0
nltk=3.8.1=pypi_0
numpy=1.23.1=pypi_0
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cuda-cupti-cu11=11.7.101=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
nvidia-cufft-cu11=10.9.0.58=pypi_0
nvidia-curand-cu11=10.2.10.91=pypi_0
nvidia-cusolver-cu11=11.4.0.1=pypi_0
nvidia-cusparse-cu11=11.7.4.91=pypi_0
nvidia-nccl-cu11=2.14.3=pypi_0
nvidia-nvtx-cu11=11.7.91=pypi_0
opencv-contrib-python=4.10.0.84=pypi_0
opencv-python=4.10.0.84=pypi_0
openssl=1.1.1w=h7f8727e_0
opt-einsum=3.4.0=pypi_0
orjson=3.11.3=pypi_0
packaging=24.2=pypi_0
pandas=2.0.3=pypi_0
pillow=9.4.0=pypi_0
pip=25.2=pyhc872135_0
platformdirs=4.4.0=pypi_0
propcache=0.3.2=pypi_0
protobuf=3.20.3=pypi_0
psutil=5.9.8=pypi_0
pthread-stubs=0.3=h0ce48e5_1
pycparser=2.23=pypi_0
pydantic=2.6.0=pypi_0
pydantic-core=2.16.1=pypi_0
pydub=0.25.1=pypi_0
pygments=2.19.2=pypi_0
pyparsing=3.1.4=pypi_0
python=3.10.0=h12debd9_5
python-dateutil=2.9.0.post0=pypi_0
python-multipart=0.0.20=pypi_0
pytorch-lightning=2.0.7=pypi_0
pytz=2024.2=pypi_0
pyyaml=6.0.2=pypi_0
readline=8.3=hc2a1206_0
regex=2024.11.6=pypi_0
requests=2.32.5=pypi_0
rich=14.1.0=pypi_0
ruff=0.13.0=pypi_0
safehttpx=0.1.6=pypi_0
safetensors=0.4.5=pypi_0
scikit-image=0.22.0=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.10.1=pypi_0
semantic-version=2.10.0=pypi_0
sentry-sdk=2.19.2=pypi_0
setproctitle=1.3.4=pypi_0
setuptools=78.1.1=py310h06a4308_0
shellingham=1.5.4=pypi_0
six=1.17.0=pypi_0
smmap=5.0.1=pypi_0
sniffio=1.3.1=pypi_0
sounddevice=0.5.1=pypi_0
spaces=0.24.2=pypi_0
sqlite=3.50.2=hb25bd0a_1
starlette=0.47.3=pypi_0
sympy=1.14.0=pypi_0
termcolor=3.1.0=pypi_0
threadpoolctl=3.5.0=pypi_0
tifffile=2025.5.10=pypi_0
tk=8.6.15=h54e0aa7_0
tokenizers=0.20.3=pypi_0
tomlkit=0.13.3=pypi_0
torch=2.0.1=pypi_0
torch-kmeans=0.2.0=pypi_0
torchmetrics=1.8.2=pypi_0
torchvision=0.15.2=pypi_0
tqdm=4.67.1=pypi_0
transformers=4.46.3=pypi_0
triton=2.0.0=pypi_0
typer=0.17.4=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024.2=pypi_0
urllib3=2.5.0=pypi_0
uvicorn=0.35.0=pypi_0
wandb=0.19.1=pypi_0
wcwidth=0.2.13=pypi_0
websockets=14.2=pypi_0
wheel=0.45.1=py310h06a4308_0
xformers=0.0.21=pypi_0
xorg-libx11=1.8.12=h9b100fa_1
xorg-libxau=1.0.12=h9b100fa_0
xorg-libxdmcp=1.1.5=h9b100fa_0
xorg-xorgproto=2024.1=h5eee18b_1
xz=5.6.4=h5eee18b_1
yarl=1.20.1=pypi_0
zipp=3.20.2=pypi_0
zlib=1.3.1=hb25bd0a_0
